pyproject.toml
[project]
name = "pokemon-discord-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "ibm-watsonx-orchestrate>=2.4.0",
    "pillow>=10.3.0",
    "prometheus-client>=0.20.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
